Fixed Collision Problem
Previously, the gems' collision box would still remain active after colliding with the player. An offset of .5 seconds is still enough for the player to hit the gems' collider a second time before its destroyed. I disabled the 'PolygonCollider' component after the first collision, not detecting a second collision if the player and the gems are within collision range.